requirements.txt
# PyTorch for CUDA 12.1 on Python 3.11/3.12
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio

# Main Application Libraries
llama-cpp-python==0.2.79
spacy==3.7.4
numpy<2.0

# NEW TTS Library
elevenlabs~=2.9.2
python-dotenv~=1.1.1